[chore] #58 패키지 재구성 - animalShelter, recommended
# Conflicts:
#	src/main/java/com/kuit/findyou/domain/information/animalShelter/dto/GetVolunteerWorksResponse.java
#	src/main/java/com/kuit/findyou/domain/information/animalShelter/dto/UpdateVolunteerWorkRequest.java
#	src/main/java/com/kuit/findyou/domain/information/controller/InformationController.java
#	src/main/java/com/kuit/findyou/domain/information/service/facade/InformationServiceFacade.java
#	src/test/java/com/kuit/findyou/global/common/util/TestInitializer.java
