+++ src/main/java/com/kuit/findyou/domain/information/animalShelter/dto/AnimalShelterPagingResponse.java
@@ -1,4 +1,4 @@
-package com.kuit.findyou.domain.information.dto;
+package com.kuit.findyou.domain.information.animalShelter.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
