+++ src/main/java/com/kuit/findyou/domain/information/animalShelter/dto/AnimalShelterResponse.java
@@ -1,4 +1,4 @@
-package com.kuit.findyou.domain.information.dto;
+package com.kuit.findyou.domain.information.animalShelter.dto;
 
 import com.kuit.findyou.domain.information.model.AnimalShelter;
 import io.swagger.v3.oas.annotations.media.Schema;
