+++ src/main/java/com/kuit/findyou/domain/information/animalShelter/repository/AnimalCenterRepository.java
@@ -1,4 +1,4 @@
-package com.kuit.findyou.domain.information.repository;
+package com.kuit.findyou.domain.information.animalShelter.repository;
 
 import com.kuit.findyou.domain.information.model.AnimalCenter;
 import org.springframework.data.jpa.repository.JpaRepository;
