+++ src/main/java/com/kuit/findyou/domain/information/animalShelter/repository/AnimalDepartmentRepository.java
@@ -1,4 +1,4 @@
-package com.kuit.findyou.domain.information.repository;
+package com.kuit.findyou.domain.information.animalShelter.repository;
 
 import com.kuit.findyou.domain.information.model.AnimalDepartment;
 import org.springframework.data.jpa.repository.JpaRepository;
