+++ src/main/java/com/kuit/findyou/domain/information/animalShelter/repository/AnimalShelterRepository.java
@@ -1,4 +1,4 @@
-package com.kuit.findyou.domain.information.repository;
+package com.kuit.findyou.domain.information.animalShelter.repository;
 
 import com.kuit.findyou.domain.information.model.AnimalShelter;
 import org.springframework.data.domain.Pageable;
