+++ src/main/java/com/kuit/findyou/domain/information/animalShelter/repository/VolunteerWorkRepository.java
@@ -1,4 +1,4 @@
-package com.kuit.findyou.domain.information.repository;
+package com.kuit.findyou.domain.information.animalShelter.repository;
 
 import com.kuit.findyou.domain.information.model.VolunteerWork;
 import org.springframework.data.domain.Pageable;
