+++ src/main/java/com/kuit/findyou/domain/information/animalShelter/service/AnimalShelterService.java
@@ -1,9 +1,7 @@
-package com.kuit.findyou.domain.information.service;
+package com.kuit.findyou.domain.information.animalShelter.service;
 
-import com.kuit.findyou.domain.information.dto.AnimalShelterPagingResponse;
-import com.kuit.findyou.domain.information.dto.AnimalShelterResponse;
-
-import java.util.List;
+import com.kuit.findyou.domain.information.animalShelter.dto.AnimalShelterPagingResponse;
+import com.kuit.findyou.domain.information.animalShelter.dto.AnimalShelterResponse;
 
 public interface AnimalShelterService {
     AnimalShelterPagingResponse<AnimalShelterResponse> getShelters(Long lastId, String type, String sido, String sigungu, Double lat, Double lng, int size);
