--- src/main/java/com/kuit/findyou/domain/information/dto/ContentType.java
@@ -1,6 +1,0 @@
-package com.kuit.findyou.domain.information.dto;
-
-public enum ContentType {
-    VIDEO,
-    NEWS
-}
