+++ src/main/java/com/kuit/findyou/domain/information/recommended/dto/ContentType.java
@@ -1,0 +1,6 @@
+package com.kuit.findyou.domain.information.recommended.dto;
+
+public enum ContentType {
+    VIDEO,
+    NEWS
+}
