+++ src/main/java/com/kuit/findyou/domain/information/recommended/dto/RecommendedContentResponse.java
@@ -1,4 +1,4 @@
-package com.kuit.findyou.domain.information.dto;
+package com.kuit.findyou.domain.information.recommended.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
