+++ src/main/java/com/kuit/findyou/domain/information/recommended/repository/RecommendedNewsRepository.java
@@ -1,6 +1,6 @@
-package com.kuit.findyou.domain.information.repository;
+package com.kuit.findyou.domain.information.recommended.repository;
 
-import com.kuit.findyou.domain.information.model.RecommendedNews;
+import com.kuit.findyou.domain.information.recommended.model.RecommendedNews;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RecommendedNewsRepository extends JpaRepository<RecommendedNews,Long> {
