+++ src/main/java/com/kuit/findyou/domain/information/recommended/repository/RecommendedVideoRepository.java
@@ -1,6 +1,6 @@
-package com.kuit.findyou.domain.information.repository;
+package com.kuit.findyou.domain.information.recommended.repository;
 
-import com.kuit.findyou.domain.information.model.RecommendedVideo;
+import com.kuit.findyou.domain.information.recommended.model.RecommendedVideo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RecommendedVideoRepository extends JpaRepository<RecommendedVideo,Long> {
