+++ src/main/java/com/kuit/findyou/domain/information/recommended/service/strategy/RecommendedContentStrategy.java
@@ -1,0 +1,9 @@
+package com.kuit.findyou.domain.information.recommended.service.strategy;
+
+import com.kuit.findyou.domain.information.recommended.dto.RecommendedContentResponse;
+
+import java.util.List;
+
+public interface RecommendedContentStrategy {
+    List<RecommendedContentResponse> getRecommendedContents();
+}
