+++ src/main/java/com/kuit/findyou/domain/information/service/facade/InformationServiceFacade.java
@@ -3,8 +3,8 @@ package com.kuit.findyou.domain.information.service.facade;
 import com.kuit.findyou.domain.information.dto.AnimalShelterPagingResponse;
 import com.kuit.findyou.domain.information.dto.AnimalShelterResponse;
 import com.kuit.findyou.domain.information.dto.ContentType;
-import com.kuit.findyou.domain.information.dto.GetVolunteerWorksResponse;
 import com.kuit.findyou.domain.information.dto.RecommendedContentResponse;
+import com.kuit.findyou.domain.information.dto.GetVolunteerWorksResponse;
 import com.kuit.findyou.domain.information.service.AnimalShelterService;
 import com.kuit.findyou.domain.information.service.RecommendedContentService;
 import com.kuit.findyou.domain.information.service.VolunteerWorkService;
